Moved routes into separate js file and imported to index. Removed border bottom cover
